chore(web): added suggestions from pr #5322
