+++ apps/web/src/pages/settings/user-profile-page/UserProfilePage.tsx
@@ -1,6 +1,8 @@
+import { FC } from 'react';
+
 import { SettingsPageContainer } from '../SettingsPageContainer';
 
-export const UserProfilePage = () => {
+export const UserProfilePage: FC = () => {
   return (
     <SettingsPageContainer title="User profile">
       <div>TODO: replace with page content</div>
